scripts/requirements.txt
numpy>=1.20.0
pillow>=9.0.0
imageio>=2.25.0
imageio-ffmpeg>=0.4.8
av>=10.0.0
